Overhauled way that grid is updated and redrawn from Qt event loop. Also now using a bunch of fixed-speed buttons. GUI remains responsive even at high update speeds thanks to special QTimer feature. Bunch of new icons to go with this new feature.

+++ src/matrix.cpp
@@ -340,8 +340,8 @@ istream& operator>>(istream& is, Matrix& m)
     list<list<float> > entries;
     is >> entries;
 
-    int cols = entries.front().size();
-    int rows = entries.size();
+    unsigned int cols = static_cast<unsigned int>(entries.front().size());
+    unsigned int rows = static_cast<unsigned int>(entries.size());
     for_iterate(r, entries)
         assert (cols == r->size());
